requirements.txt
jenkinsapi
PyYaml
pyzmq
six